evently_backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.5.0
pymongo==4.10.1
djangorestframework-simplejwt==5.5.1
Pillow==11.1.0
python-slugify==8.0.4
python-dotenv==1.0.1